Moved script tags to bottom of page
Moved the script tags to the bottom of canvas they're loaded after
canvas, and assigned them as async so they can be loaded in any time
after that.
There's an article I was reading about the best thing to do in these
situations, but it didn't cover much over dependencies on DOM elements,
except waiting for the DOM to load the element you need. Since they're
called to load after canvas, and setup() is in the file it should be
fine.
If, however, we wanted info from the entire DOM at the get go, we may
have to alter. For just canvas this should be fine.
Open to advice on how to improve/best practice.