Helping with js script loading
Moved all js into a single file (could have kept the three files and
instead organized their load sequence with some nifty JQuery, but for a
small app this works).
Declared it as async, and the script call is made just after the Canvas
element is loaded.
This makes sure we have canvas, but the page can continue to render
while we wait for any js to download.