JS loading solution and updated tests
With the application the way it is (mostly visual), there are only about
2 methods that could be made static (getting the random numbers). With
this in mind, I've put everything back into one file and have it loading
async after the DOM elements required have been parsed to load too.
If there are/were more functions required to load that were not tied
into the visuals, these could be extracted out and loaded async before
at the beginning. For now this seems the best solution. The fastest
loading of the js file, when appropriate, without holding up any later
rendering.